% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch22_individual.R
\name{CO_Embed2_Dist_tau_d_expfit_meandiff}
\alias{CO_Embed2_Dist_tau_d_expfit_meandiff}
\title{Function to calculate a statistical feature}
\usage{
CO_Embed2_Dist_tau_d_expfit_meandiff(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- CO_Embed2_Dist_tau_d_expfit_meandiff(x)

}
\author{
Trent Henderson
}
