% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DN_HistogramMode_10}
\alias{DN_HistogramMode_10}
\title{Function to calculate a statistical feature}
\usage{
DN_HistogramMode_10(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- stats::rnorm(100)
outs <- DN_HistogramMode_10(x)

}
\author{
Carl H. Lubba
}
