% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DN_Spread_Std}
\alias{DN_Spread_Std}
\title{Function to calculate a statistical feature}
\usage{
DN_Spread_Std(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- stats::rnorm(100)
outs <- DN_Spread_Std(x)

}
\author{
Trent Henderson
}
