% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_list}
\alias{feature_list}
\title{All features available in Rcatch22 in tidy format}
\format{
A vector with 1 variable:
\describe{
  \item{feature}{Name of the feature}
}
}
\usage{
feature_list
}
\description{
The variables include:
}
\keyword{datasets}
