% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PD_PeriodicityWang_th0_01}
\alias{PD_PeriodicityWang_th0_01}
\title{Function to calculate a statistical feature}
\usage{
PD_PeriodicityWang_th0_01(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- stats::rnorm(100)
outs <- PD_PeriodicityWang_th0_01(x)

}
\author{
Carl H. Lubba
}
