% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SB_BinaryStats_mean_longstretch1}
\alias{SB_BinaryStats_mean_longstretch1}
\title{Function to calculate a statistical feature}
\usage{
SB_BinaryStats_mean_longstretch1(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- stats::rnorm(100)
outs <- SB_BinaryStats_mean_longstretch1(x)

}
\author{
Carl H. Lubba
}
