% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SB_MotifThree_quantile_hh}
\alias{SB_MotifThree_quantile_hh}
\title{Function to calculate a statistical feature}
\usage{
SB_MotifThree_quantile_hh(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- 1 + 0.5 * 1:1000 + arima.sim(list(ma = 0.5), n = 1000)
outs <- SB_MotifThree_quantile_hh(x)

}
\author{
Carl H. Lubba
}
