% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SP_Summaries_welch_rect_centroid}
\alias{SP_Summaries_welch_rect_centroid}
\title{Function to calculate a statistical feature}
\usage{
SP_Summaries_welch_rect_centroid(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- stats::rnorm(100)
outs <- SP_Summaries_welch_rect_centroid(x)

}
\author{
Carl H. Lubba
}
