% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CO_Embed2_Dist_tau_d_expfit_meandiff}
\alias{CO_Embed2_Dist_tau_d_expfit_meandiff}
\title{Function to calculate a statistical feature}
\usage{
CO_Embed2_Dist_tau_d_expfit_meandiff(x)
}
\arguments{
\item{x}{a numerical time-series input vector}
}
\value{
scalar value that denotes the calculated time-series statistic
}
\description{
Function to calculate a statistical feature
}
\examples{
x <- stats::rnorm(100)
outs <- CO_Embed2_Dist_tau_d_expfit_meandiff(x)

}
\author{
Carl H. Lubba
}
